add output GeoTiff option

+++ data_cache/code/output/GFEDv4s_sp_daily.R
@@ -1,9 +1,9 @@
 # =========================================
-# GFEDv4s: HDF5 to NetCDF
+# GFEDv4s: HDF5 to NetCDF, GeoTiff
 # by species [kg m-2 s-1]
 # monthly files, daily timesteps, 0.25deg
 # =========================================
-# last updated: Feb 16, 2019
+# last updated: Feb 22, 2019
 # Tianjia Liu
 
 rm(list=ls())
@@ -15,17 +15,19 @@ source('~/Google Drive/scripts/R/fire_inv/globalParams.R')
 xYears <- 2003:2016
 xMonths <- 1:12
 varNameL <- c("CO","BC","OC","NOx","SO2")
+outputType <- "tif"
 
-GFEDv4s_nc <- function(varName, xYears, xMonths) {
+GFEDv4s_pro <- function(varName, xYears, xMonths, outputType="tif") {
   
   invName <- "GFEDv4s"
   timestamp(prefix=paste("Started...","##------ "))
   message("TASKS: ",invName," || ",varName," (",xYears[1]," to ",xYears[length(xYears)],")")
   message(paste0("/",month.abb[xMonths],"/"))
   
-  # directories to input files, output NetCDFs
+  # directories to input files, output NetCDFs, GeoTiffs
   inputFolder <- file.path(inputDir,invName)
-  outputFolder <- file.path(outputDir,invName)
+  if (outputType == "tif") {outputFolder <- file.path(outputDir_tif,invName)}
+  if (outputType == "nc") {outputFolder <- file.path(outputDir_nc,invName)}
   
   # species names
   orig_varName <- as.character(spNames[which(spNames$Name==varName),invName])
@@ -90,50 +92,68 @@ GFEDv4s_nc <- function(varName, xYears, xMonths) {
       emi_sp <- emiDM * (emiDM_SAVA*EFs_sp$SAVA + emiDM_BORF*EFs_sp$BORF + emiDM_TEMF*EFs_sp$TEMF  + 
                            emiDM_DEFO*EFs_sp$DEFO + emiDM_PEAT*EFs_sp$PEAT + emiDM_AGRI*EFs_sp$AGRI)
       
-      inv_sp <- array(0,dim=c(nLon,nLat,nDays))
+      if (outputType == "nc") {inv_sp <- array(0,dim=c(nLon,nLat,nDays))}
+      if (outputType == "tif") {inv_sp <- list()}
       for (iDay in seq_along(monthDays)) {
         # daily fraction
         emi_frac <- readGFED(prefix,paste0("emissions/",monthStr,"/daily_fraction/day_",iDay))
         
         # input species [kg/ m2/ s]
-        invDay_vec <-  emi_sp * emi_frac / (24*60*60) / 1000
-        inv_sp[,,iDay] <- matrix(as.vector(invDay_vec),nrow=nLon,ncol=nLat)[,order(unique(cellLat))]
+        invDay_ras <-  emi_sp * emi_frac / (24*60*60) / 1000
+        if (outputType == "tif") {inv_sp[[iDay]] <- invDay_ras}
+        if (outputType == "nc") {
+          inv_sp[,,iDay] <- matrix(as.vector(invDay_ras),
+                                   nrow=nLon,ncol=nLat)[,order(unique(cellLat))]
+        }
+      }
+      
+      # --------
+      # Output
+      # --------
+      subFolder <- file.path(outputFolder,xYears[iYear])
+      dir.create(subFolder,showWarnings=F,recursive=T)
+      
+      setwd(subFolder)
+      outname <- paste0(invName,"_",xYears[iYear],"_",monthStr,"_",varName)
+      
+      # -----------------
+      # Save Geotiff
+      # -----------------
+      if (outputType == "tif") {
+        writeRaster(stack(inv_sp),paste0(outname,".tif"),format="GTiff",overwrite=T)
       }
       
       # -----------------
-      # Construct NetCDF
+      # Save NetCDF
       # -----------------
-      subFolder <- file.path(outputFolder,xYears[iYear])
-      dir.create(subFolder,showWarnings=F)
-      
-      setwd(subFolder)
-      
-      # initialize NetCDF
-      ncfname <- paste0(invName,"_",xYears[iYear],"_",monthStr,"_",varName,".nc")
-      ifelse(file.exists(ncfname),file.remove(ncfname),F)
-      
-      # long, lat
-      lonDim <- ncdim_def("lon","degrees",as.double(LonVec),longname="longitude") 
-      latDim <- ncdim_def("lat","degrees",as.double(LatVec),longname="latitude") 
-      
-      # time
-      datesYrMon <- blankDates(iMonth,iMonth,xYears[iYear])
-      timeVec <- YMDtoHrsSince(datesYrMon$Year,datesYrMon$Month,datesYrMon$Day)
-      timeDim <- ncdim_def("time",paste("hours since",originDate,"00:00:00"),
-                           as.double(timeVec),longname="time")
-      
-      # input species
-      ncvar <- ncvar_def(varName,varUnits,list(lonDim,latDim,timeDim),fillval,
-                         varNameLong,compression=3)
-      
-      # add variables to NetCDF
-      ncout <- nc_create(ncfname,ncvar,force_v4=T)
-      ncvar_put(ncout,ncvar,inv_sp)
-      ncatt_put(ncout,"lon","axis","lon")
-      ncatt_put(ncout,"lat","axis","lat")
-      ncatt_put(ncout,"time","axis","time")
-      
-      nc_close(ncout)
+      if (outputType == "nc") {
+        # initialize NetCDF
+        ncfname <- paste0(outname,".nc")
+        ifelse(file.exists(ncfname),file.remove(ncfname),F)
+        
+        # long, lat
+        lonDim <- ncdim_def("lon","degrees",as.double(LonVec),longname="longitude") 
+        latDim <- ncdim_def("lat","degrees",as.double(LatVec),longname="latitude") 
+        
+        # time
+        datesYrMon <- blankDates(iMonth,iMonth,xYears[iYear])
+        timeVec <- YMDtoHrsSince(datesYrMon$Year,datesYrMon$Month,datesYrMon$Day)
+        timeDim <- ncdim_def("time",paste("hours since",originDate,"00:00:00"),
+                             as.double(timeVec),longname="time")
+        
+        # input species
+        ncvar <- ncvar_def(varName,varUnits,list(lonDim,latDim,timeDim),fillval,
+                           varNameLong,compression=3)
+        
+        # add variables to NetCDF
+        ncout <- nc_create(ncfname,ncvar,force_v4=T)
+        ncvar_put(ncout,ncvar,inv_sp)
+        ncatt_put(ncout,"lon","axis","lon")
+        ncatt_put(ncout,"lat","axis","lat")
+        ncatt_put(ncout,"time","axis","time")
+        
+        nc_close(ncout)
+      }
     }
   }
   timestamp(prefix=paste("Finished! ","##------ "))
@@ -141,7 +161,7 @@ GFEDv4s_nc <- function(varName, xYears, xMonths) {
 
 cat("\014")
 for (varName in varNameL) {
-  GFEDv4s_nc(varName, xYears, xMonths)
+  GFEDv4s_pro(varName, xYears, xMonths, outputType)
 }
 
 exit()
